refactor: Statistics 일부 기능을 위임하기 위해 Ranks 객체 생성 -Statistics -Ranks
